Move around projects, made RavenLib, made web server.
