+++ WebServer/Program.cs
@@ -1,0 +1,4 @@
+﻿using RavenLib;
+
+Server WebServer = new Server(8080);
+WebServer.Start();
